reorganizing classes for a better structure
