+++ espnff/espnff.py
@@ -2,6 +2,7 @@ import requests
 
 from .utils import (two_step_dominance,
                     power_points, )
+from .team import Team
 
 
 class ESPNFFException(Exception):
@@ -131,48 +132,6 @@ class League(object):
                     matchup.away_team = team
 
         return result
-
-
-class Team(object):
-    '''Teams are part of the league'''
-    def __init__(self, data):
-        self.team_id = data['teamId']
-        self.team_abbrev = data['teamAbbrev']
-        self.team_name = "%s %s" % (data['teamLocation'], data['teamNickname'])
-        self.division_id = data['division']['divisionId']
-        self.division_name = data['division']['divisionName']
-        self.wins = data['record']['overallWins']
-        self.losses = data['record']['overallLosses']
-        self.points_for = data['record']['pointsFor']
-        self.points_against = data['record']['pointsAgainst']
-        self.owner = "%s %s" % (data['owners'][0]['firstName'],
-                                data['owners'][0]['lastName'])
-        self.schedule = []
-        self.scores = []
-        self.mov = []
-        self._fetch_schedule(data)
-
-    def __repr__(self):
-        return 'Team(%s)' % (self.team_name, )
-
-    def _fetch_schedule(self, data):
-        '''Fetch schedule and scores for team'''
-        matchups = data['scheduleItems']
-
-        for matchup in matchups:
-            if not matchup['matchups'][0]['isBye']:
-                if matchup['matchups'][0]['awayTeamId'] == self.team_id:
-                    score = matchup['matchups'][0]['awayTeamScores'][0]
-                    opponentId = matchup['matchups'][0]['homeTeamId']
-                else:
-                    score = matchup['matchups'][0]['homeTeamScores'][0]
-                    opponentId = matchup['matchups'][0]['awayTeamId']
-            else:
-                score = matchup['matchups'][0]['homeTeamScores'][0]
-                opponentId = matchup['matchups'][0]['homeTeamId']
-
-            self.scores.append(score)
-            self.schedule.append(opponentId)
 
 
 class Settings(object):
